Issue 90 - Enable unsafe SSL/TLS renegotiation option.
This JVM property cannot be enabled on real-time, as it is established the first time the Java JSSE library is initialized, so it must be turned on before any HTTPS traffic takes place. 

The new option provides users an informational message regarding this. Two new message properties have been added to the corresponding language files.

git-svn-id: https://zaproxy.googlecode.com/svn/trunk@1187 0fcac13d-f916-6cb6-7431-acacdca54389