Issue 1215 - Alert Tree Ordering/Sorting Issue
Changed to do the ordering of the alerts with the alert node's risk which already has a lower risk (i.e. "-1") when false positive (the FPs were being shown before the INFO alerts because the latter has risk of value 0 which leads to same value that was used for FP).

git-svn-id: https://zaproxy.googlecode.com/svn/trunk@5216 0fcac13d-f916-6cb6-7431-acacdca54389